+++ src/app/(multi-footer)/[username]/[reflectionCUID]/page.client.tsx
@@ -4,7 +4,7 @@ import { Box } from "@mui/material";
 import { sqsAPI } from "@/src/api/send-to-sqs-api";
 import { animation } from "@/src/features/common/animation";
 import { ReflectionArticle } from "@/src/features/routes/reflection-detail/article";
-import ReflectionSettingHeader from "@/src/features/routes/reflection-detail/reflection-setting/ReflectionSettingHeader";
+import ReflectionSettingHeader from "@/src/features/routes/reflection-detail/header/ReflectionSettingHeader";
 import { UserInformationSection } from "@/src/features/routes/reflection-detail/user-information";
 import { useParseTagsToValue } from "@/src/hooks/reflection-tag/useParseTagsToValue";
